refactor: rename PinBottomSheet to PinDialog and reorganize PIN models
