+++ app/src/main/java/org/dhis2/usescases/general/SessionManagerActivity.kt
@@ -26,10 +26,10 @@ import org.dhis2.usescases.main.MainActivity
 import org.dhis2.usescases.qrScanner.ScanActivity
 import org.dhis2.usescases.splash.SplashActivity
 import org.dhis2.utils.analytics.AnalyticsHelper
-import org.dhis2.mobile.login.pin.ui.components.PinMode
 import org.dhis2.utils.analytics.CLICK
 import org.dhis2.utils.analytics.FORGOT_CODE
 import org.dhis2.mobile.login.pin.addPinBottomSheet
+import org.dhis2.mobile.login.pin.domain.model.PinMode
 import org.koin.android.ext.android.inject
 import javax.inject.Inject
